clk: qcom: gdsc-regulator: Add the qcom,poll-cfg-gdscr flag
The default behavior of the GDSC enable/disable sequence is
to poll the status bits of either the actual GDSCR or the
corresponding HW_CTRL registers.
On targets which have support for a CFG_GDSCR register, the
status bits might not show the correct state of the GDSC,
especially in the disable sequence, where the status bit
will be cleared even before the core is completely power
collapsed. On targets with this issue, poll the power on/off
bits in the CFG_GDSCR register instead to correctly determine
the GDSC state.

Change-Id: If41f3f1cea25c001938f28bbb94af3310860d60f
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>