clk: Add the CLK_CHILD_NO_RATE_PROP flag
The voter clock mechanism relies on the rate request
propagation to the parent clock to be one-way and not
have the clock framework propagate the rate across all
its children. Add the CLK_CHILD_NO_RATE_PROP flag for
these votable clocks to use in order to avoid the top-down
propagation of any scaling requests.

Change-Id: Ib264068797e44c7f1712aebe45318ea3e9230a1b
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
[ab123321] adapt for msm-4.14
Signed-off-by: Artem Labazov <123321artyom@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>