media: msm: vidc: Implement Venus system idle indicator exclusion
Some SoCs do feature a Venus firmware that has no system idle
indicator, meaning that the IDLE_STATUS (bit 30 of the register
VIDC_CPU_CS_SCIACMDARG0_HFI_CTRL_PC_READY in SCIACMDARG0 space,
also called HFI_CTRL_STATUS) is never set by the RTOS that is
running on the core: in this case, if we check this register,
we will (obviously) always fail power collapse, as the RTOS
will never set it (hence never declare itself "ready for power
collapse").

To address this firmware bug, rely only on the BIT 0 of this
register (WFI_STATUS), to decide if power collapse is doable.

Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>