dsp: q6voice: Force channel info command for some v2 firmwares
Some SoCs were released with a DSP firmware using CVP v2.0 and
CVD v2.3, but the same versions got a modification on later
firmwares found on, for example, SM8150.

Unfortunately, the new firmware doesn't declare a different
version for CVP nor CVD, but the old one does not support the
device_channels command, with the channel_info command being
the only option to set the CVP channels right: this is found
to happen on SDM845's DSP firmware.

For this reason, check if the current machine is SDM845 and
force sending the CVP channel info command on there: this
way, the DSP doesn't return any error and we're able to use
advanced features, like AEC and (A)ANC.

Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>