staging: ion: Add support for ion CMA heap to allocate from carveouts
Because of an XPU errata we need to temporarily have the ION CMA heap
support allocating from a careveout which isn't mapped into the kernel.
Once the errata is fixed this change can be reverted. This patch is
written with reference from commit "a568a844adc" in msm-4.14.

[Pavel Dubrova]
Adapted for 4.19 kernel. Original commit:
https://source.codeaurora.org/quic/la/kernel/msm-5.4/commit/?h=LA.UM.9.16.r1-08500-MANNAR.0&id=b97c60442af50434b493901e956bac63b4d40c04

Change-Id: Id2bbb81644be514bdb90e743a028e65a05ebd55e
Signed-off-by: Liam Mark <lmark@codeaurora.org>
Signed-off-by: Vivek Kumar <vivekuma@codeaurora.org>
Signed-off-by: Zhiqiang Tu <ztu@codeaurora.org>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>